tests/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
requests-mock==1.11.0
Pillow==10.2.0
imagehash==4.3.1 